decoupled inmemory classes
