+++ tests/integration/test_client_e2e.py
@@ -23,18 +23,19 @@ from harness_commons.engine.telemetry import TelemetryStorageConsumer, Telemetry
     TelemetryStorageProducerAsync
 from splitio.engine.impressions.manager import Counter as ImpressionsCounter
 from splitio.engine.impressions.unique_keys_tracker import UniqueKeysTracker, UniqueKeysTrackerAsync
-from splitio.events.events_delivery import EventsDelivery
-from splitio.events.events_manager import EventsManager, EventsManagerAsync
-from splitio.events.events_manager_config import EventsManagerConfig
-from splitio.events.events_task import EventsTask, EventsTaskAsync
+from harness_commons.events.events_delivery import EventsDelivery
+from harness_commons.events.events_manager import EventsManager, EventsManagerAsync
+from harness_commons.events.events_manager_config import EventsManagerConfig
+from harness_commons.events.events_task import EventsTask, EventsTaskAsync
 from splitio.models import splits
 from harness_commons.models import segments, rule_based_segments
 from harness_commons.models.events import SdkEvent
 from harness_commons.models.fallback_config import FallbackTreatmentsConfiguration, FallbackTreatmentCalculator
 from harness_commons.models.fallback_treatment import FallbackTreatment
 from splitio.recorder.recorder import StandardRecorder, PipelinedRecorder, StandardRecorderAsync, PipelinedRecorderAsync
-from splitio.storage.inmemmory import InMemoryEventStorage, InMemoryImpressionStorage, \
-    InMemorySegmentStorage, InMemorySplitStorage, InMemoryTelemetryStorage, InMemorySplitStorageAsync,\
+from splitio.storage.inmemory import InMemorySplitStorage, InMemorySplitStorageAsync
+from harness_commons.storage.inmemmory import InMemoryEventStorage, InMemoryImpressionStorage, \
+    InMemorySegmentStorage, InMemoryTelemetryStorage, \
     InMemoryEventStorageAsync, InMemoryImpressionStorageAsync, InMemorySegmentStorageAsync, \
     InMemoryTelemetryStorageAsync, InMemoryRuleBasedSegmentStorage, InMemoryRuleBasedSegmentStorageAsync 
 from splitio.storage.redis import RedisEventsStorage, RedisImpressionsStorage, \
