+++ tests/push/test_status_tracker.py
@@ -5,7 +5,7 @@ import pytest
 from splitio.push.status_tracker import PushStatusTracker, Status, PushStatusTrackerAsync
 from splitio.push.parser import ControlType, AblyError, OccupancyMessage, ControlMessage
 from harness_commons.engine.telemetry import TelemetryStorageProducer, TelemetryStorageProducerAsync
-from splitio.storage.inmemmory import InMemoryTelemetryStorage, InMemoryTelemetryStorageAsync
+from harness_commons.storage.inmemmory import InMemoryTelemetryStorage, InMemoryTelemetryStorageAsync
 from harness_commons.models.telemetry import StreamingEventTypes, SSEStreamingStatus, SSEConnectionError
