+++ tests/sync/test_events_synchronizer.py
@@ -6,7 +6,7 @@ import pytest
 
 from harness_commons.api.client import HttpResponse
 from harness_commons.api import APIException
-from splitio.storage import EventStorage
+from harness_commons.storage import EventStorage
 from harness_commons.models.events import Event
 from splitio.sync.event import EventSynchronizer, EventSynchronizerAsync
